feat: add new email template with message
